template.requirements.txt
# change to pybfe for Batfish Enterprise
pybatfish
jinja2
capirca
pyyaml
pytest
netaddr
ansible